+++ packages/radix-ui/src/components/banner.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { type HTMLAttributes, useEffect, useState } from 'react';
+import { type CSSProperties, type HTMLAttributes, useEffect, useState } from 'react';
 import { X } from 'lucide-react';
 import { cn } from '@/utils/cn';
 import { buttonVariants } from '@/components/ui/button';
@@ -131,7 +131,7 @@ function flow({ colors }: { colors: string[] }) {
             backgroundImage: `repeating-linear-gradient(70deg, ${[...colors, colors[0]].map((color, i) => `${color} ${(i * 50) / colors.length}%`).join(', ')})`,
             backgroundSize: '200% 100%',
             filter: 'saturate(2)',
-          } as object
+          } as CSSProperties
         }
       />
       <style>