+++ packages/radix-ui/src/components/codeblock.tsx
@@ -3,6 +3,7 @@ import { Check, Clipboard } from 'lucide-react';
 import {
   type ComponentProps,
   createContext,
+  type CSSProperties,
   type HTMLAttributes,
   type ReactNode,
   type RefObject,
@@ -137,7 +138,7 @@ export function CodeBlock({
               ? `line ${Number(props['data-line-numbers-start'] ?? 1) - 1}`
               : undefined,
             ...viewportProps.style,
-          } as object
+          } as CSSProperties
         }
       >
         {children}