+++ packages/radix-ui/src/layouts/notebook/slots/container.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { cn } from '@/utils/cn';
-import { useEffect, useState, type ComponentProps } from 'react';
+import { type CSSProperties, useEffect, useState, type ComponentProps } from 'react';
 import { useNotebookLayout } from '../client';
 
 export function Container(props: ComponentProps<'div'>) {
@@ -41,7 +41,7 @@ export function Container(props: ComponentProps<'div'>) {
           '--fd-docs-row-3': 'calc(var(--fd-docs-row-2) + var(--fd-toc-popover-height))',
           '--fd-sidebar-col': collapsed ? '0px' : 'var(--fd-sidebar-width)',
           ...props.style,
-        } as object
+        } as CSSProperties
       }
       className={cn(
         'grid overflow-x-clip min-h-(--fd-docs-height) auto-cols-auto auto-rows-auto [--fd-docs-height:100dvh] [--fd-header-height:0px] [--fd-toc-popover-height:0px] [--fd-sidebar-width:0px] [--fd-toc-width:0px] data-[column-changed=true]:transition-[grid-template-columns]',