+++ packages/asyncapi/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @fumadocs/asyncapi
 
+## 0.0.2
+
+### Patch Changes
+
+- 5017289: Use stable `fuma-translate`
+- Updated dependencies [5017289]
+- Updated dependencies [7a77722]
+  - @fumadocs/api-docs@0.0.2
+  - fumadocs-ui@16.10.1
+  - fumadocs-core@16.10.1
+
 ## 0.0.1
 
 ### Patch Changes