+++ packages/openapi/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @fuma-docs/openapi
 
+## 11.0.2
+
+### Patch Changes
+
+- 2b79077: fix missing legacy export
+
+## 11.0.1
+
+### Patch Changes
+
+- 5017289: Use stable `fuma-translate`
+- Updated dependencies [5017289]
+- Updated dependencies [7a77722]
+  - @fumadocs/api-docs@0.0.2
+  - fumadocs-ui@16.10.1
+  - fumadocs-core@16.10.1
+
 ## 11.0.0
 
 ### Major Changes