+++ packages/story/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @fumadocs/story
 
+## 1.1.1
+
+### Patch Changes
+
+- 5017289: Use stable `fuma-translate`
+- Updated dependencies [5017289]
+- Updated dependencies [7a77722]
+  - fumadocs-ui@16.10.1
+  - fumadocs-core@16.10.1
+
 ## 1.1.0
 
 ### Minor Changes